pyproject.toml
[tool.poetry]
name = "ttm4115-group-14"
version = "0.1.0"
description = "Web server for TTM4115"
authors = ["Lukas Neuenschwander <lukas.neu24@gmail.com>"]
readme = "README.md"
packages = [
    {include = "state_machines", from="src"},
    {include = "hermes", from="src"},
]
# sudo apt install libpq-dev

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2"
paho-mqtt = "^1.6.1"
stmpy = "^0.7.5"
psycopg2 = "^2.9.6"
appjar = "^0.94.0"

[tool.poetry.scripts]
student_machine = 'state_machines.StudentMachine:main'
teacher_machine = 'state_machines.TeacherMachine:main'
hermes = 'hermes.manage:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
